[FEAT] Support maxrows and use CSV format
Default pretty format does not escape the pipe character and, therefore,
org gets confused about the number of columns on the table. CSV format,
on the other hand, escapes commands inside the values and, therefore, it
is safer to use.

Added the maxrows parameter to be able to return longer results but the
default has been preserved to 100.